requirements.txt
langchain>=1.2.10
langchain-openai>=1.1.11
langchain-community>=0.4.1
langgraph>=1.0.10
python-dotenv>=1.2.2
langchain-mcp-adapters>=0.2.1
langchain-text-splitters>=0.0.1
streamlit>=1.28.0
google-api-python-client>=2.100.0
